pyproject.toml
[project]
name = "convo_newsletter_crew"
version = "0.1.0"
description = "convo-newsletter-crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai>=0.100.0",
    "python-dotenv",
    "pyyaml"
]

[project.scripts]
convo_newsletter_crew = "convo_newsletter_crew.main:run"
run_crew = "convo_newsletter_crew.main:run"
train = "convo_newsletter_crew.main:train"
replay = "convo_newsletter_crew.main:replay"
test = "convo_newsletter_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
